refactor: remove test bot and move functions to official bot
