--- work-discord-bot-test/.env.example
@@ -1,7 +1,0 @@
-DISCORD_TOKEN=your_discord_bot_token
-CLIENT_ID=your_discord_application_id
-GUILD_ID=your_discord_server_id
-PREMIUM_ROLE_ID=your_premium_role_id
-SUPABASE_URL=
-SUPABASE_KEY=
-SOLANA_RPC_URL=
