--- work-discord-bot-test/.gitignore
@@ -1,3 +1,0 @@
-node_modules
-dist
-.env
