--- work-discord-bot-test/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "work-discord-bot",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "dev": "tsc && node dist/index.js",
-    "build": "tsc"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "packageManager": "pnpm@10.7.0",
-  "dependencies": {
-    "@discordjs/builders": "^1.10.1",
-    "@discordjs/rest": "^2.4.3",
-    "@solana/spl-token": "^0.4.13",
-    "@solana/web3.js": "^1.98.0",
-    "@supabase/supabase-js": "^2.49.4",
-    "@types/cors": "^2.8.17",
-    "axios": "^1.8.4",
-    "cors": "^2.8.5",
-    "discord-api-types": "^0.37.119",
-    "discord.js": "^14.18.0",
-    "dotenv": "^16.4.7",
-    "express": "^5.1.0",
-    "nodemon": "^3.1.9",
-    "tweetnacl": "^1.0.3",
-    "typescript": "^5.8.3"
-  },
-  "devDependencies": {
-    "@types/express": "5.0.1"
-  }
-}
