+++ work-discord-bot/.env.example
@@ -5,3 +5,5 @@ PREMIUM_ROLE_ID=your_premium_role_id
 SUPABASE_URL=
 SUPABASE_KEY=
 SOLANA_RPC_URL=
+CLIENT_URL=
+CLIENT_TIP_URL=
